+++ app/src/test/java/reikai/novel/download/NovelDownloadManagerFailureTest.kt
@@ -83,6 +83,7 @@ class NovelDownloadManagerFailureTest {
         saver = mockk(),
         securityPreferences = SecurityPreferences(InMemoryPreferenceStore()),
         adultChecker = mockk { coEvery { adultNovelIdsAmong(any()) } returns emptySet() },
+        getNovelCategories = mockk(),
     )
 
     @BeforeEach